+++ tests/scenarios/compat-resolver-test.ts
@@ -41,7 +41,11 @@ Scenarios.fromProject(() => new Project())
     Qmodule(scenario.name, function (hooks) {
       let expectTranspiled: (file: string) => ReturnType<ReturnType<ExpectFile>['transform']>;
       let givenFiles: (files: Record<string, string>) => void;
-      let configure: (opts?: Partial<CompatResolverOptions['options']>, extraOpts?: ConfigureOpts) => Promise<void>;
+      let configure: (
+        opts?: Partial<CompatResolverOptions['options']>,
+        extraOpts?: ConfigureOpts,
+        emberVersion?: string
+      ) => Promise<void>;
 
       interface ConfigureOpts {
         appPackageRules?: Partial<PackageRules>;
@@ -59,13 +63,17 @@ Scenarios.fromProject(() => new Project())
             outputFileSync(resolve(app.dir, filename), contents, 'utf8');
           }
         };
-        configure = async function (opts?: Partial<CompatResolverOptions['options']>, extraOpts?: ConfigureOpts) {
+        configure = async function (
+          opts?: Partial<CompatResolverOptions['options']>,
+          extraOpts?: ConfigureOpts,
+          emberVersion = '4.6.0' //based on app-template package.json
+        ) {
           let etcOptions: EtcOptions = {
             compilerPath: require.resolve('ember-source-latest/dist/ember-template-compiler'),
             targetFormat: 'hbs',
             transforms: [
               ...(extraOpts?.astPlugins ?? []),
-              [require.resolve('@embroider/compat/src/resolver-transform'), { appRoot: app.dir }],
+              [require.resolve('@embroider/compat/src/resolver-transform'), { appRoot: app.dir, emberVersion }],
             ],
           };
 
@@ -1137,6 +1145,36 @@ Scenarios.fromProject(() => new Project())
             fn,
             array,
             concat
+          }),
+        });
+      `);
+      });
+
+      test('built-in helper unique-id is not imported when used with ember source version <5.2', async function () {
+        givenFiles({
+          'templates/application.hbs': `{{(unique-id)}}`,
+        });
+        await configure({ staticHelpers: true }, undefined, '4.6.0');
+        expectTranspiled('templates/application.hbs').equalsCode(`
+        import { precompileTemplate } from "@ember/template-compilation";
+        export default precompileTemplate("{{(unique-id)}}", {
+          moduleName: "my-app/templates/application.hbs",
+        });
+      `);
+      });
+
+      test('built-in helper unique-id is imported when used with ember source version >=5.2', async function () {
+        givenFiles({
+          'templates/application.hbs': `{{(unique-id)}}`,
+        });
+        await configure({ staticHelpers: true }, undefined, '5.2.0');
+        expectTranspiled('templates/application.hbs').equalsCode(`
+        import { precompileTemplate } from "@ember/template-compilation";
+        import { uniqueId } from "@ember/helper";
+        export default precompileTemplate("{{(uniqueId)}}", {
+          moduleName: "my-app/templates/application.hbs",
+          scope: () => ({
+            uniqueId,
           }),
         });
       `);